fix: Allow lookup for all types of parent tuples in Traintuple type's output Fill()
using the ledger method GetOutModelKeyChecksumAddress() to lookup each parent's key, which allows not only Traintuple types as parents but also Aggregatetuple and CompositeTraintuple types.

Code in this file for Traintuple output Fill() is now consistent with the equivalent Fill() methods of AggregateTuple's and CompositeTraintuple's respective output files.

This solves a bug where Traintuples cannot retrieve non Traintuples (Aggregatetuples or CompositeTraintuples) as parent inputs. After submitting a compute plan that contains such a Traintuple, the chaincode and thus the backend and sdk, cli or frontend lose their ability to list models. The compute plan executes fine but listing tuples is now broken.

More specifically the smart contract methods queryTraintuple() or queryTraintuples() are not being able to find the parent Aggregatetuple or CompositeTraintuple when calling Fill() on a Traintuple that has a parent Aggregatetuple or CompositeTraintuple. The tuple (model) list API reports the error  message "could not retrieve parent traintuple with key PARENT-KEY-ID - traintuple PARENT-KEY-ID not found" which it's generated by the said Fill() method.

Signed-off-by: Elias Kouskoumvekakis <eliaskousk@gmail.com>